requirements.txt
absl-py==1.4.0
appdirs @ file:///home/conda/feedstock_root/build_artifacts/appdirs_1603108395799/work
astunparse==1.6.3
blobfile==2.0.0
brotlipy @ file:///home/conda/feedstock_root/build_artifacts/brotlipy_1648854175163/work
cachetools==5.3.0
certifi==2022.12.7
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1636046063618/work
charset-normalizer==3.0.1
click @ file:///home/conda/feedstock_root/build_artifacts/click_1666798198223/work
contourpy==1.0.6
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1637687014355/work
cycler==0.11.0
docker-pycreds==0.4.0
einops==0.6.0
filelock==3.8.2
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
gitdb @ file:///home/conda/feedstock_root/build_artifacts/gitdb_1669279893622/work
GitPython @ file:///home/conda/feedstock_root/build_artifacts/gitpython_1672338156505/work
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
importlib-metadata==6.0.0
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.2
mpi4py @ file:///croot/mpi4py_1671223370575/work
numpy==1.24.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
pathtools==0.1.2
Pillow==9.3.0
protobuf==3.20.1
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pycryptodomex==3.16.0
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
pyparsing==3.0.9
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.8.2
pytz==2022.7
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1648757091578/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1673863902341/work
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.9.3
sentry-sdk @ file:///home/conda/feedstock_root/build_artifacts/sentry-sdk_1674487741133/work
setproctitle @ file:///home/conda/feedstock_root/build_artifacts/setproctitle_1649637304940/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap @ file:///home/conda/feedstock_root/build_artifacts/smmap_1611376390914/work
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX @ file:///home/conda/feedstock_root/build_artifacts/tensorboardx_1654638215170/work
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
threadpoolctl==3.1.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1665144421445/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1673452138552/work
wandb @ file:///home/conda/feedstock_root/build_artifacts/wandb_1673613476298/work
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
